Integracion del agendamiento al frontMovil - Funcional

+++ Producto/backend/src/main/java/com/nodevet/app/dto/flow/PagoInitRequestDTO.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class PagoInitRequestDTO {
     // El frontend solo nos dirá: "Quiero pagar la reserva número X"
     // El backend se encargará de buscar en la BD cuánto cuesta esa reserva.
-    private Long idReserva; 
+    private Integer idReserva; 
 }
